Add class representing CustomTextField
